v1.1.0
Added withoutZeros function to Double.  Added .backport.navigationTitle() function for older iOS.  Fixed JSON coding issue (since we're using codable, don't need to verify that all the contents are actually JSON supported NSObjects).  Added additional version tests.  Added injection tests with a count to include expected failure and run count.  Fixed so debug breakpoints are accessible from the proper thread instead of being stranded on the main thread.  Added Placard shape.  Added Triangle shaped.  Fixed .backport.background(color)

+++ Sources/UI/BytesView.swift
@@ -40,7 +40,7 @@ public struct RandomBytesTestView: View {
         List {
             ForEach(0..<10, id: \.self) { index in
                 let pow = Int(pow(10, index).description) ?? 1
-                let bytes = Int.random(max: 123456) * pow
+                let bytes = Int64(Int.random(max: 123456)) * Int64(pow)
                 BytesView(label: "\"\(bytes)\"", bytes: bytes)
             }
         }
